requirements.txt
MySQL-python
Django
djangorestframework
django-cors-headers